+++ src/main/java/io/github/mortuusars/wares/integration/kubejs/WaresKubeJSPlugin.java
@@ -1,0 +1,10 @@
+package io.github.mortuusars.wares.integration.kubejs;
+
+import dev.latvian.mods.kubejs.KubeJSPlugin;
+
+public class WaresKubeJSPlugin extends KubeJSPlugin {
+    @Override
+    public void registerEvents() {
+        WaresJSEvents.register();
+    }
+}